+++ src/main/java/com/bigcart/user/managementservice/bigcartusermanagementservice/domain/repository/GuestRepository.java
@@ -10,6 +10,6 @@ import org.springframework.data.repository.CrudRepository;
 // This will be AUTO IMPLEMENTED by Spring into a Bean called userRepository
 // CRUD refers Create, Read, Update, Delete
 
-public interface GuestRepository extends CrudRepository<Guest, Integer> {
+public interface GuestRepository extends CrudRepository<Guest, Long> {
 
 }
